add delete for escape room and interface for dbconnection
