+++ EscapeRoom/src/main/java/DAO/Parser.java
@@ -1,21 +1,20 @@
 package DAO;
 
-import connections.callback.ParsingCallBack;
-import connections.query.resultAttribute.Attribute;
+import connections.callback.ParsingCallback;
+import connections.attribute.outputAttribute.OutputAttribute;
 
 import java.util.HashSet;
-import java.util.Map;
 
 public class Parser<T> {
 
-    ParsingCallBack<T> callback;
+    ParsingCallback<T> callback;
 
-    public Parser(ParsingCallBack<T> callback) {
+    public Parser(ParsingCallback<T> callback) {
         this.callback = callback;
     }
 
-    public void parseObject(T object, HashSet<Attribute> values) {
-        for (Attribute attribute: values) {
+    public void parseObject(T object, HashSet<OutputAttribute> values) {
+        for (OutputAttribute attribute: values) {
             switch (attribute.getType()) {
                 case STRING -> callback.onCallbackString(object, attribute);
                 case INT -> callback.onCallbackInt(object, attribute);
