+++ EscapeRoom/src/main/java/connections/DbConnection.java
@@ -1,133 +1,13 @@
 package connections;
 
-import connections.query.resultAttribute.Attribute;
-import connections.query.resultAttribute.AttributeValue;
-import exceptions.ConnectionException;
-import connections.query.queryAttribute.QueryAttribute;
+import connections.attribute.outputAttribute.OutputAttribute;
+import connections.attribute.queryAttribute.QueryAttribute;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.sql.*;
-import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 
-public class DbConnection {
-    private static final String FILENAME = "src/Password.txt";
-    private static final String DRIVERCLASS = "com.mysql.cj.jdbc.Driver";
-    private static final String URL = "jdbc:mysql://localhost:3306/db-escaperoom";
-    private static final String USER = "root";
-    private String password;
-
-    private static final DbConnection dbConnection = new DbConnection();
-
-    public static DbConnection getInstance() {
-        return dbConnection;
-    }
-
-    private DbConnection() {
-        try {
-            this.password = readPassword();
-        } catch (IOException e) {
-            System.err.println("Error. Could not read the file.");
-        }
-    }
-
-    public void callCreate(String query, List<QueryAttribute> queryAttributes) {
-        try(Connection connection = this.createConnection();
-            PreparedStatement statement = this.createStatementWithQuery(connection, query)) {
-            this.addAttributesToStatement(statement, queryAttributes);
-            this.executeCreate(statement);
-        } catch (ConnectionException | SQLException e) {
-            System.out.println(e.getMessage());
-        }
-    }
-
-    public List<HashSet<Attribute>> callQuery(String query, List<QueryAttribute> queryAttributes, List<Attribute> attributes) {
-
-        List<HashSet<Attribute>> list = new ArrayList<>();
-
-        try(Connection connection = this.createConnection();
-            PreparedStatement statement = this.createStatementWithQuery(connection, query)) {
-
-            this.addAttributesToStatement(statement, queryAttributes);
-
-            try(ResultSet resultSet = this.executeQuery(statement)) {
-                while(resultSet.next()) {
-                    HashSet<Attribute> hashSet = new HashSet<>();
-                    for (Attribute attribute: attributes) {
-                        Attribute att = new Attribute(attribute.getName(), attribute.getType());
-                        switch (attribute.getType()) {
-                            case STRING -> {
-                                AttributeValue<String> value = new AttributeValue<>(resultSet.getString(attribute.getName()));
-                                att.setValue(value);
-                            }
-                            case INT -> {
-                                AttributeValue<Integer> value = new AttributeValue<>(resultSet.getInt(attribute.getName()));
-                                att.setValue(value);
-                            }
-                            case DOUBLE ->{
-                                AttributeValue<Double> value = new AttributeValue<>(resultSet.getDouble(attribute.getName()));
-                                att.setValue(value);
-                            }
-                        }
-                        hashSet.add(att);
-                    }
-                    list.add(hashSet);
-                }
-            }
-        } catch (ConnectionException | SQLException e) {
-            System.out.println(e.getMessage());
-        }
-        return list;
-    }
-
-    private void executeCreate(PreparedStatement statement) throws ConnectionException {
-        try {
-            statement.executeUpdate();
-        } catch (SQLException e) {
-            throw new ConnectionException("Error executing statement to create an object.");
-        }
-    }
-
-    private ResultSet executeQuery(PreparedStatement statement) throws ConnectionException {
-        try {
-            return statement.executeQuery();
-        } catch (SQLException e) {
-            throw new ConnectionException("Error executing statement to query into de ddbb.");
-        }
-    }
-
-    private Connection createConnection() throws ConnectionException {
-        try {
-            Class.forName(DRIVERCLASS);
-            return DriverManager.getConnection(URL, USER, this.password);
-        } catch (ClassNotFoundException | SQLException e) {
-            throw new ConnectionException("Error while attempting connection to the database.");
-        }
-    }
-
-    private PreparedStatement createStatementWithQuery(Connection connection, String query) throws ConnectionException {
-        try {
-            return connection.prepareStatement(query);
-        } catch (SQLException e) {
-            throw new ConnectionException("Error while creating statement from query " + query);
-        }
-    }
-
-    private void addAttributesToStatement(PreparedStatement statement, List<QueryAttribute> attributes) throws ConnectionException {
-        for (QueryAttribute attribute: attributes) {
-            try {
-                attribute.addToStatement(statement);
-            } catch (ConnectionException e) {
-                throw new ConnectionException(e.getMessage());
-            }
-        }
-    }
-
-    private static String readPassword() throws IOException {
-        Path fileName = Path.of(DbConnection.FILENAME);
-        return Files.readString(fileName);
-    }
+public interface DbConnection {
+    public void create(String query, List<QueryAttribute> queryAttributes);
+    public void delete(String query, List<QueryAttribute> queryAttributes);
+    public List<HashSet<OutputAttribute>> query(String query, List<QueryAttribute> queryAttributes, List<OutputAttribute> outputAttributes);
 }
