+++ EscapeRoom/src/main/java/connections/attribute/outputAttribute/OutputAttribute.java
@@ -1,14 +1,14 @@
-package connections.query.resultAttribute;
+package connections.attribute.outputAttribute;
 
 import java.util.Objects;
 
-public class Attribute {
+public class OutputAttribute {
 
     private String name;
     private ResultType type;
     private AttributeValue value;
 
-    public Attribute(String name, ResultType type) {
+    public OutputAttribute(String name, ResultType type) {
         this.name = name;
         this.type = type;
     }
@@ -33,7 +33,7 @@ public class Attribute {
     public boolean equals(Object object) {
         if (this == object) return true;
         if (object == null || getClass() != object.getClass()) return false;
-        Attribute attribute = (Attribute) object;
+        OutputAttribute attribute = (OutputAttribute) object;
         return Objects.equals(getName(), attribute.getName()) && getType() == attribute.getType() && Objects.equals(getValue(), attribute.getValue());
     }
