+++ EscapeRoom/src/main/java/connections/attribute/queryAttribute/QueryAttribute.java
@@ -1,4 +1,4 @@
-package connections.query.queryAttribute;
+package connections.attribute.queryAttribute;
 
 import exceptions.ConnectionException;
